Add separate version module
